+++ recognize.py
@@ -22,7 +22,7 @@ def define_nn_crf(name, embed_mat, seq_len, class_num):
     vocab_num, embed_len = embed_mat.shape
     embed = Embedding(input_dim=vocab_num, output_dim=embed_len,
                       weights=[embed_mat], input_length=seq_len, trainable=True)
-    input = Input(shape=(seq_len,), dtype='int32')
+    input = Input(shape=(seq_len,))
     embed_input = embed(input)
     func = map_item(name, funcs)
     crf = K_CRF(class_num)
